New call ringtone
The previous ringtone is bad and this new one is good.

Since many of Android's sounds are managed by the system, this sound
is only played while waiting for an outgoing call to connect.

The new sound is created by Ana Gelez who has graciously agreed
to distribute the ringtones under the terms of the Apache License.
https://github.com/matrix-org/matrix-react-sdk/pull/4500/files